% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R, R/SaemixRes.R
\docType{methods}
\name{[}
\alias{[}
\alias{[<-,SaemixData-method}
\alias{[,SaemixData-method}
\alias{[,SaemixRepData-method}
\alias{[<-,SaemixRepData-method}
\alias{[,SaemixSimData-method}
\alias{[<-,SaemixSimData-method}
\alias{[<-,SaemixRes-method}
\title{Get/set methods for SaemixData object}
\usage{
\S4method{[}{SaemixData}(x, i, j, drop)

\S4method{[}{SaemixRepData}(x, i, j, drop)

\S4method{[}{SaemixRepData}(x, i, j) <- value

\S4method{[}{SaemixSimData}(x, i, j, drop)

\S4method{[}{SaemixSimData}(x, i, j) <- value

\S4method{[}{SaemixRes}(x, i, j) <- value
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}

\item{value}{value to replace with}
}
\description{
Access slots of a SaemixData object using the object["slot"] format
}
\keyword{methods}
