% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\docType{methods}
\name{logLik}
\alias{logLik}
\alias{logLik.SaemixObject}
\alias{AIC.SaemixObject}
\alias{BIC.SaemixObject}
\title{Extract likelihood from an SaemixObject resulting from a call to saemix}
\usage{
\method{logLik}{SaemixObject}(object, method = c("is", "lin", "gq"), add = TRUE, ...)

\method{AIC}{SaemixObject}(object, ..., k = 2)

\method{BIC}{SaemixObject}(object, ...)
}
\arguments{
\item{object}{name of an SaemixObject object}

\item{method}{character string, one of c("is","lin","gq"), to select one of the available approximations to the log-likelihood (is: Importance Sampling; lin: linearisation and gq: Gaussian Quadrature).}

\item{add}{boolean, specifying if the requested value should be added to the object in case it is not present. The default is TRUE.}

\item{...}{additional arguments}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC}
}
\value{
Returns the selected statistical criterion (log-likelihood, AIC, BIC) extracted from the SaemixObject, computed with the 'method' argument if given.
}
\description{
The likelihood in saemix can be computed by one of three methods: linearisation (linearisation of the model), importance sampling (stochastic integration) and gaussian quadrature (numerical integration). The linearised likelihood is obtained as a byproduct of the computation of the Fisher Information Matrix (fim=TRUE in the control options given to the saemix function).
If no method argument is given, this function will attempt to extract the likelihood computed by importance sampling (method="is").
If the requested likelihood is not present in the object and the argument add=TRUE (the default), it will be computed and added to the object in addition to returning the computed value. If it is not present and add=FALSE, an error message is given.
}
\references{
Comets  E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Kuhn E, Lavielle M. Maximum likelihood estimation in nonlinear mixed effects models. Computational Statistics and Data Analysis 49, 4 (2005), 1020-1038.

Comets E, Lavenu A, Lavielle M. SAEMIX, an R version of the SAEM algorithm. 20th meeting of the Population Approach Group in Europe, Athens, Greece (2011), Abstr 2173.
}
\seealso{
\link{AIC}, \link{BIC}, \link{saemixControl}, \link{saemix}, \link{llis.saemix}], \link{llgq.saemix}, \link{fim.saemix}
}
\author{
Emmanuelle Comets \email{emmanuelle.comets@inserm.fr}

Audrey Lavenu

Marc Lavielle.
}
\keyword{methods}
