% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModel.R
\name{mydiag}
\alias{mydiag}
\title{Matrix diagonal}
\usage{
mydiag(x = 1, nrow, ncol)
}
\arguments{
\item{x}{a matrix, vector or 1D array, or missing.}

\item{nrow}{Optional number of rows for the result when x is not a matrix.}

\item{ncol}{Optional number of columns for the result when x is not a matrix.}
}
\value{
If x is a matrix then diag(x) returns the diagonal of x. The resulting vector will have names if the matrix x has matching column and rownames.
}
\description{
Extract or replace the diagonal of a matrix, or construct a diagonal matrix (replace diag function from R-base)
}
\examples{

mydiag(1)
mydiag(c(1,2))

}
\seealso{
\code{diag}
}
\author{
Emmanuelle Comets \href{mailto:emmanuelle.comets@inserm.fr}{emmanuelle.comets@inserm.fr}, Audrey Lavenu,
Marc Lavielle.
}
\keyword{models}
