% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\docType{data}
\name{oxboys.saemix}
\alias{oxboys.saemix}
\title{Heights of Boys in Oxford}
\format{
This data frame contains the following columns:
\describe{
\item{Subject}{an ordered factor giving a unique identifier for each boy in the experiment }
\item{age}{a numeric vector giving the standardized age (dimensionless)}
\item{height}{a numeric vector giving the height of the boy (cm)}
\item{Occasion}{ an ordered factor - the result of converting 'age' from a continuous variable to a count so these slightly unbalanced data can be analyzed as balanced } }
}
\usage{
oxboys.saemix
}
\description{
The \code{oxboys.saemix} data collects the height and age of students in Oxford measured over time.
The data frame has 234 rows and 4 columns.
}
\details{
These data are described in Goldstein (1987) as data on the height of a selection of boys from Oxford, England
versus a standardized age. The dataset can be found in the package \code{nlme}.
We use an linear model for this data:
y_ij = Base_i + slope_i x_ij +epsilon_ij
}
\examples{
data(oxboys.saemix)
saemix.data<-saemixData(name.data=oxboys.saemix,header=TRUE,
      name.group=c("Subject"),name.predictors=c("age"),name.response=c("height"),
      units=list(x="yr",y="cm"))

# plot the data
plot(saemix.data)

growth.linear<-function(psi,id,xidep) {
  x<-xidep[,1]
  base<-psi[id,1]
  slope<-psi[id,2]
  f<-base+slope*x
  return(f)
}
saemix.model<-saemixModel(model=growth.linear,description="Linear model",
      psi0=matrix(c(140,1),ncol=2,byrow=TRUE,dimnames=list(NULL,c("base","slope"))),
      transform.par=c(1,0),covariance.model=matrix(c(1,1,1,1),ncol=2,byrow=TRUE), 
      error.model="constant")

saemix.options<-list(algorithms=c(1,1,1),nb.chains=1,seed=201004,
      save=FALSE,save.graphs=FALSE,displayProgress=FALSE)
\donttest{
saemix.fit<-saemix(saemix.model,saemix.data,saemix.options)
}

}
\references{
JC Pinheiro, DM Bates (2000) \emph{Mixed-effects Models in S and S-PLUS}, Springer, New York (Appendix A.19)
}
\keyword{datasets}
