% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{plot-SaemixData}
\alias{plot-SaemixData}
\alias{plot,SaemixData-methods}
\alias{plot,SaemixData}
\alias{plot,SaemixSimData-method}
\alias{plot,SaemixSimData}
\title{Plot of longitudinal data}
\usage{
\S4method{plot}{SaemixData}(x, y, ...)

\S4method{plot}{SaemixSimData}(x, y, irep = -1, ...)
}
\arguments{
\item{x}{an SaemixData object or an SaemixSimData object}

\item{y}{unused, present for compatibility with base plot function}

\item{...}{additional arguments to be passed on to plot (titles, legends, ...)}

\item{irep}{number of replicate datasets to use in the mirror plot}
}
\description{
This function will plot a longitudinal dataframe contained in an SaemixData object. By default it produces a spaghetti plot, but arguments can be passed on to modify this behaviour.
}
\keyword{plot}
