% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModel.R
\name{plot.saemixModel}
\alias{plot.saemixModel}
\alias{plot.SaemixModel}
\title{Plot model predictions for a new dataset}
\usage{
plot.saemixModel(smx.model, smx.data, psi = NA)
}
\arguments{
\item{smx.model}{an SaemixModel object}

\item{smx.data}{an SaemixData object}

\item{psi}{a vector or a dataframe giving the parameters for which predictions are to be computed (defaults to NA).
The number of columns in psi (or the number of elements of psi, if psi is given as a vector) should match the number of
parameters in the model, otherwise an error message will be shown and the function will return empty.
If psi is NA, the predictions are computed for the population parameters in the model (first line of the psi0 slot).
If psi is a dataframe, each line will be used for a separate subject of the smx.data object. Elements of psi will be recycled
if psi has less lines than the number of subjects in the dataset.}
}
\value{
a ggplot object
}
\description{
Plot model predictions for a new dataset
}
\details{
The function uses the model slot of the SaemixModel object to obtain predictions, using the dataset contained in the
SaemixData object. The user is responsible for making sure data and model match.
if psi is not given, the predictions will be computed for the population parameters (first line of the psi0 slot) of the object.

Currently this function only works for models defined as 'structural'.
}
\examples{
# TODO
}
