% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixObject.R
\docType{methods}
\name{predict-methods}
\alias{predict-methods}
\alias{predict,ANY-method}
\alias{predict,SaemixObject-method}
\title{Methods for Function predict}
\usage{
\S4method{predict}{SaemixObject}(
  object,
  newdata = NULL,
  type = c("ipred", "ypred", "ppred", "icpred"),
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{an SaemixObject}

\item{newdata}{an optional dataframe for which predictions are desired. If newdata is given, it must contain the predictors needed for the model in object}

\item{type}{the type of predictions (ipred= individual, ppred=population predictions obtained with the population estimates, ypred=mean of the population predictions, icpred=conditional predictions). With newdata, individual parameters can be estimated if the new data contains observations; otherwise, predictions correspond to the population predictions ppred, and type is ignored.}

\item{se.fit}{whether the SE are to be taken into account in the model predictions}

\item{...}{additional arguments passed on to fitted()}
}
\value{
a vector or a dataframe (if more than one type) with the corresponding predictions for each observation in the dataframe
}
\description{
Methods for function \code{predict}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"ANY\")")}{Default predict functions}

\item{list("signature(object = \"SaemixObject\")")}{Computes predictions
using the results of an SAEM fit} }
}

\keyword{methods}
