% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModel.R
\name{predict.saemixmodel}
\alias{predict.saemixmodel}
\alias{predict.SaemixModel}
\title{Predictions for a new dataset}
\usage{
predict.saemixmodel(object, predictors, psi = NA, id = NA)
}
\arguments{
\item{object}{an SaemixModel object}

\item{predictors}{a dataframe with the predictors for the model (must correspond to the predictors used by the model function)}

\item{psi}{a vector or a dataframe giving the parameters for which predictions are to be computed (defaults to NA).
The number of columns in psi (or the number of elements of psi, if psi is given as a vector) should match the number of
parameters in the model, otherwise an error message will be shown and the function will return empty.
If psi is NA, the predictions are computed for the population parameters in the model (first line of the psi0 slot).
If psi is a dataframe, each line will be used for a separate 'subject' in the predictors dataframe, as
indicated by the id argument; if id is not given, only the first line of psi will be used.}

\item{id}{a vector of indices of length equal to the number of lines in predictors, matching each line of predictors to the
corresponding line in psi, ie the parameters for this predictors (defaults to NA). If id is given, the unique values in id must be equal
to the number of lines in psi, otherwise id will be set to 1. If id is given and its values do not take the consecutive values 1:N, the
indices will be matched to 1:N to follow the lines in psi.}
}
\value{
a list with two components
\describe{
\item{param}{a dataframe with the estimated parameters}
\item{predictions}{a dataframe with the population predictions}
}
}
\description{
Predictions for a new dataset
}
\details{
The function uses the model slot of the SaemixModel object to obtain predictions, using the predictors object. The
user is responsible for giving all the predictors needed by the model function.
if psi is not given, the predictions will be computed for the population parameters (first line of the psi0 slot) of the object.

The predictions correspond to the structure of the model; for models defined in terms of their likelihood, the predictions
are the log-pdf of the model (see documentation for details).

Warning: this function is currently under development and the output may change in future versions of the package to conform to the usual predict functions.
}
\examples{
# TODO
}
