% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{readSaemix,SaemixData-method}
\alias{readSaemix,SaemixData-method}
\alias{readSaemix,SaemixData}
\title{Create a longitudinal data structure from a file or a dataframe

Helper function not intended to be called by the user}
\usage{
\S4method{readSaemix}{SaemixData}(object, dat = NULL)
}
\arguments{
\item{object}{an SaemixData object}

\item{dat}{the name of a dataframe in the R environment, defaults to NULL; if NULL, the function will
attempt to read the file defined by the slot name.data.}
}
\description{
Create a longitudinal data structure from a file or a dataframe

Helper function not intended to be called by the user
}
