% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\docType{methods}
\name{readSaemix-methods}
\alias{readSaemix-methods}
\alias{readSaemix}
\title{Methods for Function read}
\usage{
readSaemix(object, dat = NULL)
}
\description{
Reads longitudinal data to create a SaemixData object (internal)
}
\keyword{internal}
