% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\name{replaceData}
\alias{replaceData}
\alias{replaceData.saemixObject}
\alias{replaceData-methods}
\title{Replace the data element in a SaemixObject object}
\usage{
replaceData.saemixObject(saemixObject, newdata)
}
\arguments{
\item{saemixObject}{an SaemixObject object}

\item{newdata}{a dataframe containing data}
}
\value{
an object of class \code{"\linkS4class{SaemixObject}"}. The population parameters are retained but all the predictions, individual parameters and statistical criteria are removed. The function attempts to extract the elements entering the statistical model (subject id, predictors, covariates and response).
}
\description{
Returns an SaemixObject object where the data object has been replaced by the data provided in a dataframe
}
\examples{
# TODO
}
\keyword{methods}
