% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/func_aux.R
\name{saemix.internal}
\alias{saemix.internal}
\alias{.First.lib}
\alias{cutoff}
\alias{cutoff.eps}
\alias{cutoff.max}
\alias{cutoff.res}
\alias{normcdf}
\alias{norminv}
\alias{error}
\alias{error.typ}
\alias{ssq}
\alias{transpsi}
\alias{transphi}
\alias{derivphi}
\alias{dtransphi}
\alias{compute.Uy}
\alias{compute.LLy}
\alias{conditional.distribution}
\alias{trnd.mlx}
\alias{gammarnd.mlx}
\alias{tpdf.mlx}
\alias{conditional.distribution_c}
\alias{conditional.distribution_d}
\title{Internal saemix objects}
\usage{
cutoff(x, seuil = .Machine$double.xmin)
}
\description{
Internal saemix objects.
}
\details{
These are not to be called by the user.
}
\keyword{internal}
