% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\name{saemix.predict}
\alias{saemix.predict}
\title{Compute model predictions after an saemix fit}
\usage{
saemix.predict(object)
}
\arguments{
\item{object}{an SaemixObject object}
}
\value{
an updated SaemixObject object
}
\description{
In nonlinear mixed effect models, different types of predictions may be obtained, including individual predictions and population predictions.
This function takes an SaemixObject and adds any missing predictions for maximum a posteriori and conditional mean estimations of the individual
parameters, and for the different types of individual and population predictions for the response variable.
}
\keyword{methods}
