% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{saemix}
\alias{saemix}
\alias{initialiseMainAlgo}
\alias{mstep}
\alias{estep}
\title{Stochastic Approximation Expectation Maximization (SAEM) algorithm}
\usage{
saemix(model, data, control = list(), population = NULL)
}
\arguments{
\item{model}{an object of class SaemixModel, created by a call to the
function \code{\link{saemixModel}}}

\item{data}{an object of class SaemixData, created by a call to the function
\code{\link{saemixData}}}

\item{control}{a list of options, see \code{\link{saemixControl}}}

\item{population}{optional population-distribution specification created by
\code{\link{copulaPopulation}}. Explicit \code{NULL} selects the ordinary
Gaussian saemix model and algorithm. For compatibility, omission uses a
preceding low-level \code{copulaSet} declaration when one is active.}
}
\value{
An object of class SaemixObject containing the results of the fit of
the data by the non-linear mixed effect model. A summary of the results is
printed out to the terminal, and, provided the appropriate options have not
been changed, numerical and graphical outputs are saved in a directory.
}
\description{
SAEM algorithm perform parameter estimation for nonlinear mixed effects
models without any approximation of the model (linearization, quadrature
approximation, . . . )
}
\details{
The SAEM algorithm is a stochastic approximation version of the standard EM
algorithm proposed by Kuhn and Lavielle (see reference). Details of the
algorithm can be found in the pdf file accompanying the package.
}
\examples{

data(theo.saemix)

saemix.data<-saemixData(name.data=theo.saemix,header=TRUE,sep=" ",na=NA,
   name.group=c("Id"),name.predictors=c("Dose","Time"),
   name.response=c("Concentration"),name.covariates=c("Weight","Sex"),
   units=list(x="hr",y="mg/L", covariates=c("kg","-")), name.X="Time")

model1cpt<-function(psi,id,xidep) { 
	  dose<-xidep[,1]
	  tim<-xidep[,2]  
	  ka<-psi[id,1]
	  V<-psi[id,2]
	  CL<-psi[id,3]
	  k<-CL/V
	  ypred<-dose*ka/(V*(ka-k))*(exp(-k*tim)-exp(-ka*tim))
	  return(ypred)
}

saemix.model<-saemixModel(model=model1cpt,modeltype="structural",
  description="One-compartment model with first-order absorption", 
  psi0=matrix(c(1.,20,0.5,0.1,0,-0.01),ncol=3, byrow=TRUE,
  dimnames=list(NULL, c("ka","V","CL"))),transform.par=c(1,1,1),
  covariate.model=matrix(c(0,1,0,0,0,0),ncol=3,byrow=TRUE),fixed.estim=c(1,1,1),
  covariance.model=matrix(c(1,0,0,0,1,0,0,0,1),ncol=3,byrow=TRUE),
  omega.init=matrix(c(1,0,0,0,1,0,0,0,1),ncol=3,byrow=TRUE),error.model="constant")

\donttest{
# Not run (strict time constraints for CRAN)
 saemix.fit<-saemix(saemix.model,saemix.data,list(seed=632545,directory="newtheo",
 save=FALSE,save.graphs=FALSE, print=FALSE))
 
# Prints a summary of the results
print(saemix.fit)

# Outputs the estimates of individual parameters
psi(saemix.fit)

# Shows some diagnostic plots to evaluate the fit
plot(saemix.fit)
}
}
\references{
E Comets, A Lavenu, M Lavielle M (2017). Parameter estimation in nonlinear mixed effect models using saemix,
an R implementation of the SAEM algorithm. Journal of Statistical Software, 80(3):1-41.

E Kuhn, M Lavielle (2005). Maximum likelihood estimation in nonlinear mixed effects models.
Computational Statistics and Data Analysis, 49(4):1020-1038.

E Comets, A Lavenu, M Lavielle (2011). SAEMIX, an R version of the SAEM algorithm. 20th meeting of the
Population Approach Group in Europe, Athens, Greece, Abstr 2173.
}
\seealso{
\code{\link{SaemixData}},\code{\link{SaemixModel}},
\code{\link{SaemixObject}}, \code{\link{saemixControl}},
\code{\link{plot.saemix}}
}
\author{
Emmanuelle Comets \email{emmanuelle.comets@inserm.fr}

Audrey Lavenu

Marc Lavielle
}
\keyword{models}
