% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\docType{data}
\name{toenail.saemix}
\alias{toenail.saemix}
\title{Toenail data}
\format{
This data frame contains the following columns:
\describe{
\item{id}{subject index in file}
\item{time}{time of measurement (in months)}
\item{y}{degree of onycholysis (0 if none or mild, 1 if moderate or severe)}
\item{treatment}{treatment indicator (1=Treatment A, 0= Treatment B)}
\item{visit}{visit number (visit numbers 1-7 correspond to scheduled visits at 0, 4, 8, 12, 24, 36, and 48 weeks)}
}
#'
}
\source{
prLogistic package in R
}
\usage{
toenail.saemix
}
\description{
The \code{toenail.saemix} data are from a multicenter study comparing two oral treatments for toe-nail infection, including information for 294 patients measured at 7 weeks, comprising a total of 1908 measurements. The outcome binary variable "onycholysis" indicates the degree of separation of the nail plate from the nail-bed (none or mild versus moderate or severe). Patients were evaluated at baseline (week 0) and at weeks 4, 8, 12, 24, 36, and 48 thereafter.
}
\details{
The data in the \code{toenail.saemix} was copied from the Toenail dataset provided by the prLogistic package. Different models
and analyses have been performed to describe this dataset in Molenberg and Verbeke (2000).
}
\references{
De Backer M, De Vroey C, Lesaffre E, Scheys I, De Keyser P (1998).
Twelve weeks of continuous oral therapy for toenail onychomycosis caused by dermatophytes:
A double-blind comparative trial of terbinafine 250 mg/day versus itraconazole 200 mg/day.
Journal of the American Academy of Dermatology, 38, 57-63.

Lesaffre E, Spiessens B (2001). On the effect of the number of quadrature points in a logistic random-effects model: An example.
Journal of the Royal Statistical Society, Series C, 50, 325-335.

Verbeke G, Molenberghs G (2000). Linear mixed models for longitudinal data, Springer, New York.

Rabe-Hesketh S, Skrondal A (2008). Multilevel and Longitudinal Modeling Using Stata. Mahwah, NJ: Lawrence Erlbaum Associates. Second Edition.

#' @examples
data(toenail.saemix)

#' @keywords datasets
}
