% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{transform}
\alias{transform}
\alias{transform.numeric}
\title{Transform covariates}
\usage{
\method{transform}{numeric}(
  `_data`,
  transformation = function(x) x,
  centering = "median",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{_data}{a vector with values of type numeric}

\item{transformation}{transformation function. Defaults to no transformation}

\item{centering}{string, giving the value used to center the covariate; can be "mean" or "median", in which case this value will be computed from the data, 'none' or 0 for no centering, or a value given by the user. Defaults to the median value over the dataset.}

\item{verbose}{a boolean, prints messages during the execution of the function if TRUE. Defaults to FALSE.}

\item{\dots}{unused, for consistency with the generic method}
}
\value{
a vector
}
\description{
Transform and/or center a vector
}
\examples{
# TODO
}
\keyword{data}
