% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{validate.names}
\alias{validate.names}
\title{Name validation (## )Helper function not intended to be called by the user)}
\arguments{
\item{usernames}{vector of strings}

\item{datanames}{vector of strings}

\item{recognisednames}{vector of strings, values for automatic recognition}

\item{verbose}{logical, whether to print warning messages}
}
\value{
a vector with valid names
}
\description{
Helper function, checks if the names given by the user match to the names in the dataset. If not, automatic recognition is attempted.
}
\examples{
# TODO
}
\keyword{methods}
