% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixRes.R
\name{vcov}
\alias{vcov}
\alias{vcov.SaemixRes}
\alias{vcov.SaemixObject}
\alias{##vcov,SaemixRes}
\alias{vcov,SaemixObject}
\title{Extracts the Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{SaemixRes}(object, ...)

\method{vcov}{SaemixObject}(object, ...)
}
\arguments{
\item{object}{a fitted object from a call to saemix}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
A matrix of the estimated covariances between the parameter estimates in model. In saemix, this matrix is obtained as the inverse of the Fisher Information Matrix computed by linearisation
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted model object
}
