% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixData.R
\docType{methods}
\name{read-methods}
\alias{read-methods}
\alias{read}
\alias{read,SaemixData-method}
\alias{read,SaemixData}
\title{Methods for Function read}
\usage{
read(object, verbose = TRUE)

\S4method{read}{SaemixData}(object)
}
\arguments{
\item{object}{an SaemixData object}
}
\description{
Reads longitudinal data to create a SaemixData object (internal)
}
\keyword{internal}
